multiple plots, colorers, saving
I made the last error bar example.
I made a section on multiple plots, which consists of a histogram with
a line, as well as a best fit line to noisy data.
I added a short section on colorbars, that needs to be expanded.
A section on saving plots now exists, too.